@@ -4,18 +4,18 @@ import '../redux/redux.dart';
 import '../redux_component/redux_component.dart';
 
 /// abstract for custom extends
-abstract class Adapter<T> extends Logic<T> implements AbstractAdapter<T> {
+abstract class Adapter<T> extends Logic<T?> implements AbstractAdapter<T> {
   final AdapterBuilder<T> _adapter;
 
   AdapterBuilder<T> get protectedAdapter => _adapter;
 
   Adapter({
-    @required AdapterBuilder<T> adapter,
-    Reducer<T> reducer,
-    ReducerFilter<T> filter,
-    Effect<T> effect,
-    Dependencies<T> dependencies,
-    @deprecated Object Function(T) key,
+    required AdapterBuilder<T> adapter,
+    Reducer<T?>? reducer,
+    ReducerFilter<T?>? filter,
+    Effect<T?>? effect,
+    Dependencies<T>? dependencies,
+    @deprecated Object Function(T?)? key,
   })  : assert(adapter != null),
         assert(dependencies?.adapter == null,
             'Unexpected dependencies.list for Adapter.'),
@@ -30,25 +30,25 @@ abstract class Adapter<T> extends Logic<T> implements AbstractAdapter<T> {
         );
 
   @override
-  ListAdapter buildAdapter(ContextSys<T> ctx) =>
-      ctx.enhancer
+  ListAdapter buildAdapter(ContextSys<T>? ctx) =>
+      ctx!.enhancer
           ?.adapterEnhance(protectedAdapter, this, ctx.store)
           ?.call(ctx.state, ctx.dispatch, ctx) ??
       protectedAdapter?.call(ctx.state, ctx.dispatch, ctx);
 
   @override
-  ContextSys<T> createContext(
-    Store<Object> store,
-    BuildContext buildContext,
-    Get<T> getState, {
-    @required Enhancer<Object> enhancer,
-    @required DispatchBus bus,
+  ContextSys<T?> createContext(
+    Store<Object?> store,
+    BuildContext? buildContext,
+    Get<T?> getState, {
+    required Enhancer<Object?> enhancer,
+    required DispatchBus bus,
   }) {
     assert(bus != null && enhancer != null);
-    return AdapterContext<T>(
+    return AdapterContext<T?>(
       logic: this,
       store: store,
-      buildContext: buildContext,
+      buildContext: buildContext!,
       getState: getState,
       bus: bus,
       enhancer: enhancer,
@@ -58,12 +58,12 @@ abstract class Adapter<T> extends Logic<T> implements AbstractAdapter<T> {
 
 class AdapterContext<T> extends LogicContext<T> {
   AdapterContext({
-    @required AbstractAdapter<T> logic,
-    @required Store<Object> store,
-    @required BuildContext buildContext,
-    @required Get<T> getState,
-    @required DispatchBus bus,
-    @required Enhancer<Object> enhancer,
+    required AbstractAdapter<T> logic,
+    required Store<Object?> store,
+    required BuildContext buildContext,
+    required Get<T> getState,
+    required DispatchBus bus,
+    required Enhancer<Object?> enhancer,
   })  : assert(bus != null && enhancer != null),
         super(
           logic: logic,
@@ -76,7 +76,7 @@ class AdapterContext<T> extends LogicContext<T> {
 
   @override
   ListAdapter buildAdapter() {
-    final AbstractAdapter<T> curLogic = logic;
+    final AbstractAdapter<T?> curLogic = logic as AbstractAdapter<T?>;
     return curLogic.buildAdapter(this);
   }
 }
