@@ -3,7 +3,7 @@ import 'package:flutter/foundation.dart';
 import '../redux/redux.dart';
 import 'op_mixin.dart';
 
-bool _listEquals(List<dynamic> list1, List<dynamic> list2) {
+bool _listEquals(List<dynamic> list1, List<dynamic>? list2) {
   if (identical(list1, list2)) {
     return true;
   }
@@ -28,18 +28,18 @@ bool _listEquals(List<dynamic> list1, List<dynamic> list2) {
   return true;
 }
 
-abstract class _BasicReselect<T, P> extends MutableConn<T, P>
+abstract class _BasicReselect<T, P> extends MutableConn<T, P?>
     with ConnOpMixin<T, P> {
-  List<dynamic> _subsCache;
-  P _pCache;
+  List<dynamic>? _subsCache;
+  P? _pCache;
   bool _hasBeenCalled = false;
 
-  List<dynamic> getSubs(T state);
+  List<dynamic> getSubs(T? state);
 
-  P reduceSubs(List<dynamic> list);
+  P reduceSubs(List<dynamic>? list);
 
   @override
-  P get(T state) {
+  P? get(T? state) {
     final List<dynamic> subs = getSubs(state);
     if (!_hasBeenCalled || !_listEquals(subs, _subsCache)) {
       _subsCache = subs;
@@ -50,55 +50,55 @@ abstract class _BasicReselect<T, P> extends MutableConn<T, P>
   }
 }
 
-abstract class Reselect1<T, P, K0> extends _BasicReselect<T, P> {
-  K0 getSub0(T state);
+abstract class Reselect1<T, P, K0> extends _BasicReselect<T, P?> {
+  K0 getSub0(T? state);
   P computed(K0 state);
 
   @override
-  List<dynamic> getSubs(T state) => <dynamic>[getSub0(state)];
+  List<dynamic> getSubs(T? state) => <dynamic>[getSub0(state)];
 
   @override
-  P reduceSubs(List<dynamic> list) => Function.apply(computed, list);
+  P? reduceSubs(List<dynamic>? list) => Function.apply(computed, list);
 }
 
-abstract class Reselect2<T, P, K0, K1> extends _BasicReselect<T, P> {
-  K0 getSub0(T state);
-  K1 getSub1(T state);
+abstract class Reselect2<T, P, K0, K1> extends _BasicReselect<T, P?> {
+  K0 getSub0(T? state);
+  K1 getSub1(T? state);
   P computed(K0 sub0, K1 sub1);
 
   @override
-  List<dynamic> getSubs(T state) => <dynamic>[getSub0(state), getSub1(state)];
+  List<dynamic> getSubs(T? state) => <dynamic>[getSub0(state), getSub1(state)];
 
   @override
-  P reduceSubs(List<dynamic> list) => Function.apply(computed, list);
+  P? reduceSubs(List<dynamic>? list) => Function.apply(computed, list);
 }
 
-abstract class Reselect3<T, P, K0, K1, K2> extends _BasicReselect<T, P> {
-  K0 getSub0(T state);
-  K1 getSub1(T state);
-  K2 getSub2(T state);
+abstract class Reselect3<T, P, K0, K1, K2> extends _BasicReselect<T, P?> {
+  K0 getSub0(T? state);
+  K1 getSub1(T? state);
+  K2 getSub2(T? state);
   P computed(K0 sub0, K1 sub1, K2 sub2);
 
   @override
-  List<dynamic> getSubs(T state) => <dynamic>[
+  List<dynamic> getSubs(T? state) => <dynamic>[
         getSub0(state),
         getSub1(state),
         getSub2(state),
       ];
 
   @override
-  P reduceSubs(List<dynamic> list) => Function.apply(computed, list);
+  P? reduceSubs(List<dynamic>? list) => Function.apply(computed, list);
 }
 
-abstract class Reselect4<T, P, K0, K1, K2, K3> extends _BasicReselect<T, P> {
-  K0 getSub0(T state);
-  K1 getSub1(T state);
-  K2 getSub2(T state);
-  K3 getSub3(T state);
+abstract class Reselect4<T, P, K0, K1, K2, K3> extends _BasicReselect<T, P?> {
+  K0 getSub0(T? state);
+  K1 getSub1(T? state);
+  K2 getSub2(T? state);
+  K3 getSub3(T? state);
   P computed(K0 sub0, K1 sub1, K2 sub2, K3 sub3);
 
   @override
-  List<dynamic> getSubs(T state) => <dynamic>[
+  List<dynamic> getSubs(T? state) => <dynamic>[
         getSub0(state),
         getSub1(state),
         getSub2(state),
@@ -106,20 +106,20 @@ abstract class Reselect4<T, P, K0, K1, K2, K3> extends _BasicReselect<T, P> {
       ];
 
   @override
-  P reduceSubs(List<dynamic> list) => Function.apply(computed, list);
+  P? reduceSubs(List<dynamic>? list) => Function.apply(computed, list);
 }
 
 abstract class Reselect5<T, P, K0, K1, K2, K3, K4>
-    extends _BasicReselect<T, P> {
-  K0 getSub0(T state);
-  K1 getSub1(T state);
-  K2 getSub2(T state);
-  K3 getSub3(T state);
-  K4 getSub4(T state);
+    extends _BasicReselect<T, P?> {
+  K0 getSub0(T? state);
+  K1 getSub1(T? state);
+  K2 getSub2(T? state);
+  K3 getSub3(T? state);
+  K4 getSub4(T? state);
   P computed(K0 sub0, K1 sub1, K2 sub2, K3 sub3, K4 sub4);
 
   @override
-  List<dynamic> getSubs(T state) => <dynamic>[
+  List<dynamic> getSubs(T? state) => <dynamic>[
         getSub0(state),
         getSub1(state),
         getSub2(state),
@@ -128,21 +128,21 @@ abstract class Reselect5<T, P, K0, K1, K2, K3, K4>
       ];
 
   @override
-  P reduceSubs(List<dynamic> list) => Function.apply(computed, list);
+  P? reduceSubs(List<dynamic>? list) => Function.apply(computed, list);
 }
 
 abstract class Reselect6<T, P, K0, K1, K2, K3, K4, K5>
-    extends _BasicReselect<T, P> {
-  K0 getSub0(T state);
-  K1 getSub1(T state);
-  K2 getSub2(T state);
-  K3 getSub3(T state);
-  K4 getSub4(T state);
-  K5 getSub5(T state);
+    extends _BasicReselect<T, P?> {
+  K0 getSub0(T? state);
+  K1 getSub1(T? state);
+  K2 getSub2(T? state);
+  K3 getSub3(T? state);
+  K4 getSub4(T? state);
+  K5 getSub5(T? state);
   P computed(K0 sub0, K1 sub1, K2 sub2, K3 sub3, K4 sub4, K5 sub5);
 
   @override
-  List<dynamic> getSubs(T state) => <dynamic>[
+  List<dynamic> getSubs(T? state) => <dynamic>[
         getSub0(state),
         getSub1(state),
         getSub2(state),
@@ -152,14 +152,14 @@ abstract class Reselect6<T, P, K0, K1, K2, K3, K4, K5>
       ];
 
   @override
-  P reduceSubs(List<dynamic> list) => Function.apply(computed, list);
+  P? reduceSubs(List<dynamic>? list) => Function.apply(computed, list);
 }
 
-abstract class Reselect<T, P> extends _BasicReselect<T, P> {
+abstract class Reselect<T, P> extends _BasicReselect<T, P?> {
   P computed(List<dynamic> list);
 
   @override
-  P reduceSubs(List<dynamic> list) => Function.apply(computed, list);
+  P? reduceSubs(List<dynamic>? list) => Function.apply(computed, list);
 }
 
 /// issue [https://github.com/alibaba/fish-redux/issues/482]
