@@ -26,15 +26,15 @@ const ApplyLikeEnhancer ApplyLikeEnhancerIdentity = _identity;
 class AOP {
   final ApplyLikeEnhancer _enhancer;
 
-  AOP(List<ApplyLikeEnhancer> enhances)
+  AOP(List<ApplyLikeEnhancer>? enhances)
       : _enhancer = enhances?.isNotEmpty == true
-            ? enhances.reduce(_combine)
+            ? enhances!.reduce(_combine)
             : ApplyLikeEnhancerIdentity;
 
-  TypedApplyLike<R> enhance<R>(Function functor) {
+  TypedApplyLike<R?>? enhance<R>(Function functor) {
     /// cast functor to ApplyLike
     final ApplyLike init = (List<dynamic> positionalArguments,
-            [Map<Symbol, dynamic> namedArguments]) =>
+            [Map<Symbol, dynamic>? namedArguments]) =>
         Function.apply(functor, positionalArguments, namedArguments);
 
     /// enhance ApplyLike
@@ -47,54 +47,54 @@ class AOP {
 
     /// cast ApplyLike to TypedApplyLike<R>
     return (List<dynamic> positionalArguments,
-        [Map<Symbol, dynamic> namedArguments]) {
-      final R result = enhanced(positionalArguments);
+        [Map<Symbol, dynamic>? namedArguments]) {
+      final R? result = enhanced(positionalArguments);
       return result;
     };
   }
 
-  R Function() withZero<R>(R Function() f) {
-    final TypedApplyLike<R> enhanced = enhance<R>(f);
+  R? Function() withZero<R>(R Function() f) {
+    final TypedApplyLike<R?>? enhanced = enhance<R>(f);
     return enhanced != null ? () => enhanced(<dynamic>[]) : f;
   }
 
-  R Function(P) withOne<R, P>(R Function(P) f) {
-    final TypedApplyLike<R> enhanced = enhance<R>(f);
+  R? Function(P) withOne<R, P>(R Function(P) f) {
+    final TypedApplyLike<R?>? enhanced = enhance<R>(f);
     return enhanced != null ? (P p) => enhanced(<dynamic>[p]) : f;
   }
 
-  R Function(P0, P1) withTwo<R, P0, P1>(R Function(P0, P1) f) {
-    final R Function(List<dynamic>) enhanced = enhance<R>(f);
+  R? Function(P0, P1) withTwo<R, P0, P1>(R Function(P0, P1) f) {
+    final R? Function(List<dynamic>)? enhanced = enhance<R>(f);
     return enhanced != null ? (P0 p0, P1 p1) => enhanced(<dynamic>[p0, p1]) : f;
   }
 
-  R Function(P0, P1, P2) withThree<R, P0, P1, P2>(R Function(P0, P1, P2) f) {
-    final TypedApplyLike<R> enhanced = enhance<R>(f);
+  R? Function(P0, P1, P2) withThree<R, P0, P1, P2>(R Function(P0, P1, P2) f) {
+    final TypedApplyLike<R?>? enhanced = enhance<R>(f);
     return enhanced != null
         ? (P0 p0, P1 p1, P2 p2) => enhanced(<dynamic>[p0, p1, p2])
         : f;
   }
 
-  R Function(P0, P1, P2, P3) withFour<R, P0, P1, P2, P3>(
+  R? Function(P0, P1, P2, P3) withFour<R, P0, P1, P2, P3>(
       R Function(P0, P1, P2, P3) f) {
-    final TypedApplyLike<R> enhanced = enhance<R>(f);
+    final TypedApplyLike<R?>? enhanced = enhance<R>(f);
     return enhanced != null
         ? (P0 p0, P1 p1, P2 p2, P3 p3) => enhanced(<dynamic>[p0, p1, p2, p3])
         : f;
   }
 
-  R Function(P0, P1, P2, P3, P4) withFive<R, P0, P1, P2, P3, P4>(
+  R? Function(P0, P1, P2, P3, P4) withFive<R, P0, P1, P2, P3, P4>(
       R Function(P0, P1, P2, P3, P4) f) {
-    final TypedApplyLike<R> enhanced = enhance<R>(f);
+    final TypedApplyLike<R?>? enhanced = enhance<R>(f);
     return enhanced != null
         ? (P0 p0, P1 p1, P2 p2, P3 p3, P4 p4) =>
             enhanced(<dynamic>[p0, p1, p2, p3, p4])
         : f;
   }
 
-  R Function(P0, P1, P2, P3, P4, P5) withSix<R, P0, P1, P2, P3, P4, P5>(
-      R Function(P0, P1, P2, P3, P4, P5) f) {
-    final TypedApplyLike<R> enhanced = enhance<R>(f);
+  R? Function(P0, P1, P2, P3, P4, P5) withSix<R, P0, P1, P2, P3, P4, P5>(
+      R Function(P0, P1, P2, P3, P4, P5?) f) {
+    final TypedApplyLike<R?>? enhanced = enhance<R>(f);
     return enhanced != null
         ? (P0 p0, P1 p1, P2 p2, P3 p3, P4 p4, P5 p5) =>
             enhanced(<dynamic>[p0, p1, p2, p3, p4, p5])
