@@ -8,14 +8,14 @@ ApplyLikeEnhancer logAOP(String tag) {
   return isDebug()
       ? (dynamic Function(List<dynamic>) functor) {
           return (List<dynamic> positionalArguments,
-              [Map<Symbol, dynamic> namedArguments]) {
+              [Map<Symbol, dynamic>? namedArguments]) {
             print('$tag input: $positionalArguments');
             final Object result = functor(positionalArguments);
             if (result is Future) {
               result.then((Object r) {
                 print('$tag output <Future>: $r');
                 return r;
-              });
+              } as FutureOr<_> Function(dynamic));
             } else {
               print('$tag output: $result');
             }
