@@ -15,7 +15,7 @@ mixin _BatchNotify<T> on Store<T> {
   void setupBatch() {
     if (!_isSetupBatch) {
       _isSetupBatch = true;
-      super.subscribe(_batch);
+      super.subscribe!(_batch);
 
       subscribe = (void Function() callback) {
         assert(callback != null);
@@ -46,7 +46,7 @@ mixin _BatchNotify<T> on Store<T> {
         });
       }
     } else {
-      final T curState = getState();
+      final T curState = getState!();
       if (!identical(_prevState, curState)) {
         _prevState = curState;
 
@@ -76,12 +76,12 @@ class _BatchStore<T> extends Store<T> with _BatchNotify<T> {
   }
 }
 
-Store<T> createBatchStore<T>(
+Store<T?> createBatchStore<T>(
   T preloadedState,
-  Reducer<T> reducer, {
-  StoreEnhancer<T> storeEnhancer,
+  Reducer<T?>? reducer, {
+  StoreEnhancer<T?>? storeEnhancer,
 }) =>
-    _BatchStore<T>(
+    _BatchStore<T?>(
       createStore(
         preloadedState,
         _appendUpdateStateReducer<T>(reducer),
@@ -94,8 +94,8 @@ Store<T> createBatchStore<T>(
 enum _UpdateState { Assign }
 
 // replace current state
-Reducer<T> _appendUpdateStateReducer<T>(Reducer<T> reducer) =>
-    (T state, Action action) => action.type == _UpdateState.Assign
+Reducer<T?> _appendUpdateStateReducer<T>(Reducer<T?>? reducer) =>
+    (T? state, Action action) => action.type == _UpdateState.Assign
         ? action.payload
         : reducer == null ? state : reducer(state, action);
 
@@ -105,22 +105,22 @@ Store<T> connectStores<T, K>(
   T Function(T, K) update,
 ) {
   final void Function() subscriber = () {
-    final T prevT = mainStore.getState();
-    final T nextT = update(prevT, extraStore.getState());
+    final T prevT = mainStore.getState!();
+    final T nextT = update(prevT, extraStore.getState!());
     if (nextT != null && !identical(prevT, nextT)) {
-      mainStore.dispatch(Action(_UpdateState.Assign, payload: nextT));
+      mainStore.dispatch!(Action(_UpdateState.Assign, payload: nextT));
     }
   };
 
-  final void Function() unsubscribe = extraStore.subscribe(subscriber);
+  final void Function() unsubscribe = extraStore.subscribe!(subscriber);
 
   /// should triggle once
   subscriber();
 
-  final Future<dynamic> Function() superMainTD = mainStore.teardown;
+  final Future<dynamic> Function()? superMainTD = mainStore.teardown;
   mainStore.teardown = () {
     unsubscribe?.call();
-    return superMainTD();
+    return superMainTD!();
   };
 
   return mainStore;
