@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:collection/collection.dart' show IterableExtension;
 import 'package:flutter/widgets.dart' hide Action, Page;
 
 import '../redux/basic.dart';
@@ -8,23 +9,23 @@ import 'basic.dart';
 import 'lifecycle.dart';
 
 AdapterBuilder<T> asAdapter<T>(ViewBuilder<T> view) {
-  return (T unstableState, Dispatch dispatch, ViewService service) {
-    final ContextSys<T> ctx = service;
+  return (T unstableState, Dispatch dispatch, ViewService? service) {
+    final ContextSys<T>? ctx = service as ContextSys<T>?;
     return ListAdapter(
       (BuildContext buildContext, int index) =>
-          view(ctx.state, dispatch, service),
+          view(ctx!.state, dispatch, service),
       1,
     );
   };
 }
 
-Reducer<T> mergeReducers<T extends K, K>(Reducer<K> sup, [Reducer<T> sub]) {
+Reducer<T> mergeReducers<T extends K, K>(Reducer<K> sup, [Reducer<T>? sub]) {
   return (T state, Action action) {
-    return sub?.call(sup(state, action), action) ?? sup(state, action);
+    return sub?.call(sup(state, action) as T, action) ?? sup(state, action) as T;
   };
 }
 
-Effect<T> mergeEffects<T extends K, K>(Effect<K> sup, [Effect<T> sub]) {
+Effect<T> mergeEffects<T extends K, K>(Effect<K> sup, [Effect<T>? sub]) {
   return (Action action, Context<T> ctx) {
     return sub?.call(action, ctx) ?? sup.call(action, ctx);
   };
@@ -32,19 +33,18 @@ Effect<T> mergeEffects<T extends K, K>(Effect<K> sup, [Effect<T> sub]) {
 
 /// combine & as
 /// for action.type which override it's == operator
-Reducer<T> asReducer<T>(Map<Object, Reducer<T>> map) => (map == null ||
+Reducer<T>? asReducer<T>(Map<Object, Reducer<T>> map) => (map == null ||
         map.isEmpty)
     ? null
     : (T state, Action action) =>
         map.entries
-            .firstWhere(
+            .firstWhereOrNull(
                 (MapEntry<Object, Reducer<T>> entry) =>
-                    action.type == entry.key,
-                orElse: () => null)
+                    action.type == entry.key)
             ?.value(state, action) ??
         state;
 
-Reducer<T> filterReducer<T>(Reducer<T> reducer, ReducerFilter<T> filter) {
+Reducer<T>? filterReducer<T>(Reducer<T>? reducer, ReducerFilter<T>? filter) {
   return (reducer == null || filter == null)
       ? reducer
       : (T state, Action action) {
@@ -58,15 +58,14 @@ typedef SubEffect<T> = FutureOr<void> Function(Action action, Context<T> ctx);
 
 /// for action.type which override it's == operator
 /// return [UserEffecr]
-Effect<T> combineEffects<T>(Map<Object, SubEffect<T>> map) =>
+Effect<T>? combineEffects<T>(Map<Object, SubEffect<T>> map) =>
     (map == null || map.isEmpty)
         ? null
         : (Action action, Context<T> ctx) {
-            final SubEffect<T> subEffect = map.entries
-                .firstWhere(
+            final SubEffect<T>? subEffect = map.entries
+                .firstWhereOrNull(
                   (MapEntry<Object, SubEffect<T>> entry) =>
                       action.type == entry.key,
-                  orElse: () => null,
                 )
                 ?.value;
 
@@ -84,9 +83,9 @@ Effect<T> combineEffects<T>(Map<Object, SubEffect<T>> map) =>
           };
 
 /// return [EffectDispatch]
-Dispatch createEffectDispatch<T>(Effect<T> userEffect, Context<T> ctx) {
+Dispatch createEffectDispatch<T>(Effect<T>? userEffect, Context<T> ctx) {
   return (Action action) {
-    final Object result = userEffect?.call(action, ctx);
+    final Object? result = userEffect?.call(action, ctx);
 
     //skip-lifecycle-actions
     if (action.type is Lifecycle && (result == null || result == false)) {
@@ -100,13 +99,13 @@ Dispatch createEffectDispatch<T>(Effect<T> userEffect, Context<T> ctx) {
 /// return [NextDispatch]
 Dispatch createNextDispatch<T>(ContextSys<T> ctx) => (Action action) {
       ctx.broadcastEffect(action);
-      ctx.store.dispatch(action);
+      ctx.store.dispatch!(action);
     };
 
 /// return [Dispatch]
-Dispatch createDispatch<T>(Dispatch onEffect, Dispatch next, Context<T> ctx) =>
+Dispatch createDispatch<T>(Dispatch? onEffect, Dispatch next, Context<T> ctx) =>
     (Action action) {
-      final Object result = onEffect?.call(action);
+      final Object? result = onEffect?.call(action);
       if (result == null || result == false) {
         next(action);
       }
@@ -114,10 +113,10 @@ Dispatch createDispatch<T>(Dispatch onEffect, Dispatch next, Context<T> ctx) =>
       return result == _SUB_EFFECT_RETURN_NULL ? null : result;
     };
 
-ViewMiddleware<T> mergeViewMiddleware<T>(List<ViewMiddleware<T>> middleware) {
+ViewMiddleware<T>? mergeViewMiddleware<T>(List<ViewMiddleware<T>> middleware) {
   return Collections.reduce<ViewMiddleware<T>>(middleware,
       (ViewMiddleware<T> first, ViewMiddleware<T> second) {
-    return (AbstractComponent<dynamic> component, Store<T> store) {
+    return (AbstractComponent<dynamic> component, Store<T>? store) {
       final Composable<ViewBuilder<dynamic>> inner = first(component, store);
       final Composable<ViewBuilder<dynamic>> outer = second(component, store);
       return (ViewBuilder<dynamic> view) {
@@ -127,7 +126,7 @@ ViewMiddleware<T> mergeViewMiddleware<T>(List<ViewMiddleware<T>> middleware) {
   });
 }
 
-AdapterMiddleware<T> mergeAdapterMiddleware<T>(
+AdapterMiddleware<T>? mergeAdapterMiddleware<T>(
     List<AdapterMiddleware<T>> middleware) {
   return Collections.reduce<AdapterMiddleware<T>>(middleware,
       (AdapterMiddleware<T> first, AdapterMiddleware<T> second) {
@@ -142,14 +141,14 @@ AdapterMiddleware<T> mergeAdapterMiddleware<T>(
   });
 }
 
-EffectMiddleware<T> mergeEffectMiddleware<T>(
+EffectMiddleware<T>? mergeEffectMiddleware<T>(
     List<EffectMiddleware<T>> middleware) {
   return Collections.reduce<EffectMiddleware<T>>(middleware,
       (EffectMiddleware<T> first, EffectMiddleware<T> second) {
     return (AbstractLogic<dynamic> logic, Store<T> store) {
-      final Composable<Effect<dynamic>> inner = first(logic, store);
-      final Composable<Effect<dynamic>> outer = second(logic, store);
-      return (Effect<dynamic> effect) {
+      final Composable<Effect<dynamic>?> inner = first(logic, store);
+      final Composable<Effect<dynamic>?> outer = second(logic, store);
+      return (Effect<dynamic>? effect) {
         return outer(inner(effect));
       };
     };
