@@ -3,18 +3,18 @@ import '../redux/redux.dart';
 import 'op_mixin.dart';
 
 class ConnOp<T, P> extends MutableConn<T, P> with ConnOpMixin<T, P> {
-  final P Function(T) _getter;
-  final void Function(T, P) _setter;
+  final P Function(T?)? _getter;
+  final void Function(T?, P)? _setter;
 
   const ConnOp({
-    P Function(T) get,
-    void Function(T, P) set,
+    P Function(T?)? get,
+    void Function(T?, P)? set,
   })  : _getter = get,
         _setter = set;
 
   @override
-  P get(T state) => _getter(state);
+  P get(T? state) => _getter!(state);
 
   @override
-  void set(T state, P subState) => _setter(state, subState);
+  void set(T? state, P subState) => _setter!(state, subState);
 }
